chore(middleware-retry): rename defaultStrategy to StandardRetryStrategy
