+++ packages/middleware-retry/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./retryMiddleware";
 export * from "./omitRetryHeadersMiddleware";
-export * from "./defaultStrategy";
+export * from "./StandardRetryStrategy";
 export * from "./configurations";
 export * from "./delayDecider";
 export * from "./retryDecider";
